+++ app/views/Condominio/visualizar.php
@@ -20,11 +20,13 @@ $setores = $data['setores'];
                    class="btn btn-default btn-circle btn-lg">
                     <i class="fa fa-list"></i>
                 </a>
-                <a href="Condominio/confirmDelete/<?php echo $condominio['cd_condominio']; ?>"
-                   data-toggle="tooltip" data-placement="top" title="Deletar!"
-                   class="btn btn-warning btn-circle btn-lg">
-                    <i class="fa fa-trash-o"></i>
-                </a>
+                    <?php if ($userDados['nivel'] == 1): ?>
+                        <a href="Condominio/confirmDelete/<?php echo $condominio['cd_condominio']; ?>"
+                           data-toggle="tooltip" data-placement="top" title="Deletar!"
+                           class="btn btn-warning btn-circle btn-lg">
+                            <i class="fa fa-trash-o"></i>
+                        </a>
+                    <?php endif; ?>
             </span>
                 <!-- Fim Botões de administração -->
             </h3>
